# Speed up `optimize_projection` by improving `is_projection_unnecessary`